+++ app/src/test/java/com/aowen/monolith/fakes/repo/FakeOmedaCityBuildRepository.kt
@@ -18,7 +18,7 @@ class FakeOmedaCityBuildRepository : BuildRepository {
         name: String?,
         role: String?,
         order: String?,
-        heroId: Int?,
+        heroId: Long?,
         skillOrder: Int?,
         currentVersion: Int?,
         modules: Int?,
